third_party/blink/web_tests/virtual/document-isolation-policy/external/wpt/html/document-isolation-policy/credentialless-dedicated-worker.https.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] fetch same-origin + credentialless worker
  assert_equals: coep:credentialless =>  expected "same_origin" but got "Worker blocked"
[FAIL] fetch same-origin + require_corp worker
  assert_equals: coep:credentialless =>  expected "same_origin" but got "Worker blocked"
[FAIL] fetch cross-origin + credentialless worker
  assert_equals: coep:none =>  expected (undefined) undefined but got (string) "cross_origin"
[FAIL] fetch cross-origin + require_corp worker
  assert_equals: coep:credentialless =>  expected "cross_origin" but got "Worker blocked"
Harness: the test ran to completion.
